move `depth_first_search` to retworkx-core + create a new traversal module together with `dfs_edges` function
